warning fix

+++ src/vm/runner.h
@@ -881,7 +881,7 @@ class ClusterJob {
     }
 
     bool AddRunnerToTask(Runner* runner, const int32_t id) {
-        if (id < 0 || id >= tasks_.size()) {
+        if (id < 0 || id >= static_cast<int32_t>(tasks_.size())) {
             LOG(WARNING) << "fail update task: task " << id << " not exist";
             return false;
         }
@@ -901,7 +901,7 @@ class ClusterJob {
             return;
         }
         for (size_t i = 0; i < tasks_.size(); i++) {
-            if (main_task_id_ == i) {
+            if (main_task_id_ == static_cast<int32_t>(i)) {
                 output << "MAIN TASK ID " << i;
             } else {
                 output << "TASK ID " << i;
